move Post entity to a higher level
