+++ demo/src/main/java/io/reist/sandbox/app/SandboxModule.java
@@ -33,6 +33,10 @@ import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
+import io.reist.sandbox.app.model.Post;
+import io.reist.sandbox.app.model.PostStorIOSQLiteDeleteResolver;
+import io.reist.sandbox.app.model.PostStorIOSQLiteGetResolver;
+import io.reist.sandbox.app.model.PostStorIOSQLitePutResolver;
 import io.reist.sandbox.app.model.Repo;
 import io.reist.sandbox.app.model.RepoStorIOSQLiteDeleteResolver;
 import io.reist.sandbox.app.model.User;
@@ -43,10 +47,6 @@ import io.reist.sandbox.app.model.local.DbOpenHelper;
 import io.reist.sandbox.app.model.remote.NestedFieldNameAdapter;
 import io.reist.sandbox.app.model.remote.SandboxApi;
 import io.reist.sandbox.feed.FeedModule;
-import io.reist.sandbox.feed.model.local.Post;
-import io.reist.sandbox.feed.model.local.PostStorIOSQLiteDeleteResolver;
-import io.reist.sandbox.feed.model.local.PostStorIOSQLiteGetResolver;
-import io.reist.sandbox.feed.model.local.PostStorIOSQLitePutResolver;
 import io.reist.sandbox.repos.ReposModule;
 import io.reist.sandbox.repos.model.local.RepoGetResolver;
 import io.reist.sandbox.repos.model.local.RepoPutResolver;
