+++ demo/src/main/java/io/reist/sandbox/feed/model/CachedFeedService.java
@@ -21,7 +21,7 @@
 package io.reist.sandbox.feed.model;
 
 import io.reist.sandbox.app.model.CachedService;
-import io.reist.sandbox.feed.model.local.Post;
+import io.reist.sandbox.app.model.Post;
 
 /**
  * Created by 4xes on 7/11/16.
