+++ demo/src/main/java/io/reist/sandbox/feed/view/FeedListFragment.java
@@ -30,10 +30,10 @@ import javax.inject.Inject;
 
 import butterknife.BindView;
 import io.reist.sandbox.R;
+import io.reist.sandbox.app.model.Post;
 import io.reist.sandbox.app.model.SandboxError;
 import io.reist.sandbox.app.view.BaseFragment;
 import io.reist.sandbox.app.view.widget.LoaderView;
-import io.reist.sandbox.feed.model.local.Post;
 import io.reist.sandbox.feed.presenter.FeedListAdapter;
 import io.reist.sandbox.feed.presenter.FeedListPresenter;
