+++ demo/src/main/java/io/reist/sandbox/feed/view/FeedListView.java
@@ -2,8 +2,8 @@ package io.reist.sandbox.feed.view;
 
 import java.util.List;
 
+import io.reist.sandbox.app.model.Post;
 import io.reist.sandbox.app.model.SandboxError;
-import io.reist.sandbox.feed.model.local.Post;
 import io.reist.sandbox.feed.presenter.FeedListPresenter;
 import io.reist.visum.view.VisumView;
