+++ demo/src/test/java/io/reist/sandbox/feed/model/FeedServiceTest.java
@@ -22,11 +22,11 @@ import dagger.Component;
 import io.reist.sandbox.BuildConfig;
 import io.reist.sandbox.app.SandboxApplication;
 import io.reist.sandbox.app.SandboxModule;
+import io.reist.sandbox.app.model.Post;
 import io.reist.sandbox.app.model.SandboxResponse;
 import io.reist.sandbox.core.RobolectricTestCase;
 import io.reist.sandbox.core.RobolectricTestRunner;
 import io.reist.sandbox.feed.FeedModule;
-import io.reist.sandbox.feed.model.local.Post;
 import io.reist.sandbox.feed.model.local.StorIoPostService;
 import io.reist.sandbox.feed.model.remote.FeedServerApi;
 import io.reist.sandbox.feed.model.remote.RetrofitFeedService;
